Flytt lagring av anmodningsperioder ut av saksflyt. Legg til endepunkt for svar på anmodnignsperiode.
